requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.10.0
pymupdf==1.21.1
nltk==3.8.1
python-dotenv==1.0.0
